Make prepare hook smarter to match issue-id with different formats
This is now able to match branches with the following formats in a case
insensitive manner:
- "username/issue-id-the-title"
- "username/issue-id/the-title"
- "issue-id/the-title"
- "issue-id-the-title"

Where `issue-id` is in the format (case insensitive): `ABC-1234`. This
means, a set of letters and numbers separated by a dash.

This also restricts the script to run on Bash versions > 4 so we can use
the `^^` to uppercase the output format.

---

Change-type: feature